My-rose-bot/requirements.txt
python-telegram-bot
motor
dnspython
flask
